1. Added hardcoded Collisions on GridBorders, 2. Added rendering to the particles, so combined they look like snow, 3. physic 1-1 rendering decoupled, gafferongames.com/game-physics/fix-your-timestep/, 4. decoupled gridborders so that they have their own shader 5. gridrendering is too slow for 32000 particles+ discarding is still too expensive + doesnt work anymore

+++ src/math3d.h
@@ -98,7 +98,7 @@ struct Vector3i{
     int x;
     int y;
     int z;
-    Vector3i(float x, float y, float z){
+    Vector3i(int x, int y, int z){
         this->x=x;
         this->y=y;
         this->z=z;
